+++ src/char_rec_nn.rs
@@ -72,7 +72,7 @@ pub fn run_prediction(image_tensor: &Tensor) -> Result<()> {
 
   // recognize character
   let instant = Instant::now();
-  let res = net.forward(&image_tensor).softmax(-1, Kind::Float);
+  let res = net.forward(&image_tensor).softmax(-1, Kind::Double);
   let (predicted_value, probability) = utils::topk(&res, 1)[0];
 
   debug!(
